Add files via upload
Utilized Circuitverse to simulate a CPU featuring an 8-bit Arithmetic Logic Unit and incorporated an 8-bit instruction code decoder for interpreting instructions and operands.
It was implemented with a Common Bus architecture with a separate Control Bus. 
Developed an instruction set comprising 13 distinct types of instructions. 
Additionally, implemented a scrolling display mechanism for the output to store historical data, ensuring the accuracy and functionality of the CPU simulation and instruction decoder while maintaining a clear and organized display of information.